requirements.txt
google-api-python-client==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==0.21.0
requests==2.26.0